pyproject.toml
[project]
name = "foss4g-2025-geofm"
version = "0.1.0"
description = "FOSS4G 2025 Demo: Loading geospatial data via odc-stac into TerraTorch for foundation model workflows"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "odc-stac>=0.3.0",
    "terratorch>=0.1.0", 
    "xarray>=2022.12.0",
    "pystac-client>=0.7.0",
    "rasterio>=1.3.0",
    "torch>=2.0.0",
    "lightning>=2.0.0",
    "numpy>=1.21.0",
    "dask[complete]>=2023.1.0",
    "geopandas>=0.12.0",
    "matplotlib>=3.5.0",
    "cartopy>=0.21.0",
    "planetary-computer>=1.0.0",
    "stackstac>=0.4.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0"
]

[project.urls]
Homepage = "https://github.com/whatnick/foss4g_2025_geofm"
Repository = "https://github.com/whatnick/foss4g_2025_geofm"
Documentation = "https://github.com/whatnick/foss4g_2025_geofm/blob/main/README.md"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
